refactor(mitm-addon): split catalog cache tests by contract
